+++ Assets/AirConsole/examples/translations/TranslationExampleLogic.cs
@@ -31,8 +31,9 @@ namespace NDream.AirConsole.Examples {
             //in this case, we are translating a greeting message that includes the master controller's nickname
             //'welcome_nickname' has an English language value of "Welcome %nickname%, glad you're here!"
             translationExample.text = AirConsole.instance.GetTranslation("welcome_nickname",
-                new Dictionary<string, string>
-                    { { "nickname", AirConsole.instance.GetNickname(AirConsole.instance.GetMasterControllerDeviceId()) } });
+                new Dictionary<string, string> {
+                    { "nickname", AirConsole.instance.GetNickname(AirConsole.instance.GetMasterControllerDeviceId()) }
+                });
 
             /* Note: to really understand how translations work, we recommend creating your own project in the AirConsole Developer Console.
              * in this example, we show how to retrieve translations, but you will not be able to view or edit translatable phrases for the