+++ Assets/AirConsole/scripts/Runtime/Plugin/Android/AndroidUnityUtils.cs
@@ -23,21 +23,18 @@ namespace NDream.AirConsole.Android.Plugin {
 #if !UNITY_ANDROID || UNITY_EDITOR
             return defaultValue;
 #endif
-            try
-            {
-                using (AndroidJavaClass unityPlayer = new("com.unity3d.player.UnityPlayer"))
-                {
+            try {
+                using (AndroidJavaClass unityPlayer = new("com.unity3d.player.UnityPlayer")) {
                     AndroidJavaObject currentActivity = unityPlayer.GetStatic<AndroidJavaObject>("currentActivity");
-                    AndroidJavaObject intent = currentActivity.Call<AndroidJavaObject>("getIntent"); 
+                    AndroidJavaObject intent = currentActivity.Call<AndroidJavaObject>("getIntent");
                     return intent.Call<bool>("getBooleanExtra", key, defaultValue);
                 }
-            }
-            catch (System.Exception e)
-            {
+            } catch (System.Exception e) {
                 AirConsoleLogger.LogWarning(() => "Error getting intent extra: " + e);
                 return defaultValue;
             }
         }
     }
+
     // ReSharper enable RedundantUsingDirective
 }